docs: extend default capabilities with cmp ones
The `default_capabilities()` function does only return the
`textDocument.completion` subset of capabilities. The correct approach
to advertise the full set of nvim-cmp's capabilities to the LSP servers
is to first get the default set of capabilities, and then extend it with
the subset returned by the function. Use this approach in the
documentation.